+++ ubo_app/main.py
@@ -15,6 +15,9 @@ dotenv.load_dotenv(Path(__file__).parent / '.dev.env')
 
 def main() -> None:
     """Instantiate the `MenuApp` and run it."""
+    # This should be imported early to set the custom loader
+    from ubo_app.load_services import load_services
+
     os.environ['KIVY_NO_CONFIG'] = '1'
     os.environ['KIVY_NO_FILELOG'] = '1'
     os.environ['KIVY_NO_CONSOLELOG'] = '1'
@@ -24,7 +27,7 @@ def main() -> None:
     setup_logging()
     setup_error_handling()
 
-    from ubo_app.utils.loop import start_event_loop
+    from ubo_app.service import start_event_loop
 
     start_event_loop()
 
@@ -43,7 +46,6 @@ def main() -> None:
         },
     )
 
-    from ubo_app.load_services import load_services
     from ubo_app.logging import logger
     from ubo_app.menu_app.menu import MenuApp